+++ src/simmate/calculators/vasp/outputs/__init__.py
@@ -1,3 +1,4 @@
 # -*- coding: utf-8 -*-
 
 from .oszicar import Oszicar
+from .vasprun import Vasprun