+++ src/simmate/database/base_data_types/forces.py
@@ -76,7 +76,7 @@ class Forces(DatabaseTable):
     @classmethod
     def _from_toolkit(
         cls,
-        structure: ToolkitStructure,
+        structure: ToolkitStructure = None,
         site_forces=None,
         lattice_stress=None,
         as_dict=False,
@@ -85,6 +85,7 @@ class Forces(DatabaseTable):
         Given site forces and lattice stress, this function builds the rest of
         the required fields for this class as a dictionary.
         """
+
         # TODO: in the future, this should accept an IonicStep toolkit object
         # or maybe Structure + Forces toolkit objects.
         site_data = (
@@ -96,7 +97,9 @@ class Forces(DatabaseTable):
                 site_forces_norm=float(numpy.linalg.norm(site_forces)),
                 site_forces_norm_per_atom=float(
                     numpy.linalg.norm(site_forces) / structure.num_sites
-                ),
+                )
+                if structure
+                else None,
             )
             if site_forces
             else {}
@@ -108,7 +111,9 @@ class Forces(DatabaseTable):
                 lattice_stress_norm=float(numpy.linalg.norm(lattice_stress)),
                 lattice_stress_norm_per_atom=float(
                     numpy.linalg.norm(lattice_stress) / structure.num_sites
-                ),
+                )
+                if structure
+                else None,
             )
             if lattice_stress
             else {}