+++ CHANGELOG.md
@@ -26,10 +26,13 @@ There is one key exception to the rules above -- and that is with `MAJOR`=0 rele
 - REST API fields can now be specified directly with the `api_filters` attribute of any `DatabaseTable` class & fields from mix-ins are automatically added
 - add `archive_fields` attribute that sets the "raw data" for the database table & fields from mix-ins are automatically added
 - accept `TOML` input files in addition to `YAML`
+- convergence plots and extras are now written for many workflow types (such as relaxations)
+- when `use_database=True`, output files are automatically written and the workup method is directly paired with the database table.
 
 **Refactors**
 - the `website.core_components.filters` module has been absorbed into the `DatabaseTable` class/module
 - yaml input for custom workflows now matches the python input format
+- workup methods are largely depreciated and now database entries are returned when a workflow has `use_database=True`
 
 **Fixes**
 - fix bug in windows dev env where `simmate run-server` fails to find python path